Fix scroll bug
Fixed a scrollTo bug. Now it should hopefully work correctly all the
time!

Added new properties for the Game class:

Game:gameWidth
Game:gameHeight

Those will get your the width and height of the stage in classes that
are not MovieClips (but need to know these things) as well as
MovieClips that know about Game but have not been themselves added to
the stage.